pyproject.toml
[tool.poetry]
name = "uniquode"
version = "2.0"
description = "Uniquode.IO Website v2"
authors = ["David Nugent <davidn@uniquode.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
psycopg2-binary = "^2.9.1"
django-redis = "^5.0.0"
django-sitetree = "^1.17.0"
django-sass = "^1.1.0"
django-markdownx = "^3.0.1"
Pygments = "^2.10.0"
markdown-markup-emoji = "^0.2.2"
mdx-emoticons = {git = "https://github.com/Uniquode/python-markdown-emoticons.git", rev = "master"}
django-simple-history = "^3.0.0"
django-taggit = "^1.5.1"
django-taggit-anywhere = "^0.1.4"
django-taggit-helpers = "^0.1.4"
django-taggit-labels = "^0.7.0"
django-taggit-templatetags = "^0.2.5"
django-env = {git = "https://github.com/apnarm/django-env.git", rev = "main"}
djangorestframework = "^3.12.4"
django-npm = {git = "https://github.com/Uniquode/django-npm.git", rev = "master"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-pythonpath = "^0.7.3"
pytest-django = "^4.4.0"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
python_paths = [
    "app",
]
env = [
    "DJANGO_ADMIN_ENABLED=true"
]
DJANGO_SETTINGS_MODULE = "conf.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
